services/data_scout_service/requirements.txt
sqlalchemy
psycopg2-binary
clickhouse-connect
pyyaml
schedule